+++ AlbertUnruhUtils/utils/__init__.py
@@ -1,2 +1,3 @@
+from .decorator import *
 from .logger import *
 from .version import *